PALMO + GS_RANKED + POLAR_PRECISION BUGFIX
The polar iterative solver gives up after 120 iterations, and sets u = alpha*E_static. When it does this, it doesn't reset atom->ef_induced_change which is used in the palmo calculation. This is typically no a problem, since convergence failure usually means a highly repulsive system. Strangely, when GS_RANKED is also turned on, sometimes you'll get a highly attractive polar energy instead. Thus, the simulation blows up. This is no longer an issue, as atom->ef_induced_change is now set to zero (aka palmo is turned off) when a convergence failure occurs.

Also, now when a convergence failure is detected but the MC move is accepted anyways, a warning is printed to stderr. The reason for the warning is because this is obviously not a good thing, but I'm not convinced that it's bad either. It may be correct to auto-reject any time the dipoles fail to converge, but this may lead to biased rejection of certain geometries. Then again, when we accept these moves, the polar energy calculated isn't very representative, since it's based on u=alpha*E_i, rather than u=alpha(E_i+E_s). Fortunately, in the test I've performed it seems that configurations with divergent dipoles correspond to those that are in very bad configurations: ones with overlapping molecules and such.

Note that this issue never *seems* to occur with polar_max_iter, since you wouldn't know if your dipole is diverging (since you'll only do a fixed number of iterations). Again, it probably isn't a big deal, since configs with divergent dipoles aren't likely to have acceptable RD energies.