MPMC no longer works with PDB files, but rather PQR coordinate files. The reason for this major change is this...
For very large simulation cells (>= 200.0 angstroms in length), the required space needed for printing individual 
coordinates is exceeded. This is the fault of the PDB file format, which is a very old standard. An evolution of 
that file format is the PQR format. Don't fret! This format is 99% the same as PDB, with the major difference 
being that columns are parsed by whitespace and not fixed lengths. This gives us the flexibility to simulate 
systems as large as is computationally feasible, along with the ability to more precisely locate atomic positions. 

This change means that a few keywords have changed, in particular:
    1.  pdb_input --> pqr_input
    2.  pdb_output --> pqr_output
    3.  pdb_restart --> pqr_restart
    4.  read_pdb_box --> read_pqr_box

Adjusting to this change could not be easier! Simply rename all .pdb files in the job directory to .pqr and you're 
all set. MPMC will now ONLY read in and write out .pqr files. Put simply, EVERYTHING that was PDB is now PQR.

Some other notable changes:
    1.  I have changed the vmdfixn.pl script to output a file called "output.pqr" instead of "output.pdb"
    2a. Extended output will only occur if any of the basis vectors specified in the input file are greater than 
        or equal to 200.0. At this point, 6 decimal places are printed to all output files.
    2b. There is currently no way to view trajectory files directly when the extended output is turned on. In order
        to view it, you must fix it in post-production. A fix will be issued in the near future.
    3.  VMD will not display the connectivity of the BOX that is written to the restart file (this is a PQR thing).
        This can easily be forced, however, by including the "-pdb" flag when executing VMD via command line:
            [ccioce@dirac]$ vmd -pdb n2.final.pqr
        Additionally, the -pdb flag is now REQUIRED when passing trajectory files into VMD:
            [ccioce@dirac]$ vmd -pdb n2.traj.pqr

Still left To Do:
    1. Add support for direct trajectory viewing in VMD when extended output is activated.
    2. Currently, extended output is turned on by reading the input file's basis vectors, which are static. This 
       needs to be a dynamic quantity, however. See /io/output.c for the code that checks this.

- Chris