Changed signature of memnullcheck() from  memnullcheck(void *, int, int) to memnullcheck( void *, int, int, char * )
The last parameter of the original was formerly a code indicating the calling 
function. This has been replaced by two parameters which will indicate the line 
in source code where the memory allocation occurred, as well as a string 
indicating the name of the source code file. These values are easily passed using 
the standard __LINE__ and __FILE__ macros. __LINE__ is replaced by the line number 
in source code where the macro appears. Similarly, __FILE__ is replaced by a
static c-string indicating the name of the source file in which the macro appears. 
Since these parameters never have to change, but are in some sense dynamic, this 
will prevent awkward parent-code ordering or renumbering if additional 
memnullcheck()'s have to be added to a function at a later time.

I have replaced all the previous calls to memnullcheck with something like the
following:

int *ptr = malloc( sizeof(int)*100 );
memnullcheck( ptr, sizeof(int)*100, __LINE__-1, __FILE__);

where __LINE__-1 indicates that the actual allocation was attempted in the 
previous line.


=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=


IF YOU COMPILE WITH GCC, memnullcheck() will now also output a stack trace so
that you can immediately see the parent function, the grandparent function, etc.
However, these will be given as hex address offsets unless you link with the 
gcc option:

-rdynamic

in which case the actual function names will be listed. The documentation says
this flag will only work on systems using GNU ld, whatever that means. I checked 
it out on CIRCE and it works... so I guess CIRCE uses GNU ld. 

I changed the configure file in the build directory to include the -rdynamic 
option (placed inside the variable $EXTRA_LIBS) for gcc, for machines 
circe & "gdb".

-bt