parallel tempering added + other changes
1. parallel tempering added. to use, invoke the following commands

parallel_tempering on
ptemp_freq 10
max_temperature 100

this will perform a temper move every 10 steps, with a max bath temperature of 200 spaced in geometric intervals
(e.g. if temperature is set to 50, and 4 cores are used then baths will be set to 50, 63, 79 and 100 K)

A few other notes on parallel tempering:
a) observables reported reflect averages taken on the coolest bath         
b) observables for all baths will be written to energy log
c) nodestats are collected over all nodes (AR and BF) 
d) tempering moves are not included in BF and AR calculations                                    
e) number of baths is dictated by the number of cores used in the calculation
f) not compatible with temperature-dependent potentials (e.g. feynman-hibbs), but IS compatible with temperature-dependent fugacities (e.g.$

2. corrected double counting error in nodestats which led to errors in BF and AR calculations  

3. changed error calculation for all observables to reflect the standard deviation of the mean     

4. fixed bug in matrix construction for polar_wolf_full

5. mpmc now writes multiple files for field and dipole outputs when running in parallel

6. core_temp and kinetic_temp are now separately written to energy log (kinetic_temp is only used in NVE calculations)

*** to do ***
add aggregation of sorbate statistics for multi-core jobs