Added the ability to visualize the surface construction routine (geometrically, not energetically). Note, documentation on how to do this is provided in the newly created "doc" folder. We really should be documenting stuffs...
To accomplish this, specify "surf_output [FILE]" in the input file (NOTE: ensemble surf must be used, this is obviously incompatible, and useless, in other ensembles). Controlling the amount of output (more or less geometries printed to a trajectory file) is accomplished by setting "surf_print_level [1-6] (default=3)" as well, in the input file. 

Lastly, this trajectory file that is output is cryptic, and must be decoded via the newly added python script (surface_traj_decode.py). This can be found in the /tools/ directory.

- CRC