Changes the way traj/output/restart files are written.
Files are now written one at a time when running in parallel (utilizing MPI_Barrier)
to prevent file i/o congestion on very highly parallel runs.

In addition, the second most-recent restart.pdb will now be backed up to restart.pdb.last
to prevent losing a restart when the code is terminated between the call to fopen and
actually performing the write.