Corrected matrices in ewald calculations that got messed up in v120 (I managed to confuse myself).
We want reciprocal_basis[p][q] * a[q] when a[q] is a real-space vector and we want to multiply it by a recip BASIS vector.

We want reciprocal_basis[q][p] * l[q] when l[q] is a set of three numbers which specifies how many times we want each recip BASIS vector to contribute when constructing a reciprocal LATTICE vector 


With this change + changes in v120, we fix the histogramming issue, the issue of placing atoms in the incorrect sim box image, and pass all the tests mentioned in rev119's write up (including energy scaling).