This allows for the Monte Carlo calculation of equilibrium population of nuclear spin isomers in uVT and NVT ensembles if quantum rotation calculations are active.  Most of this code has been latent for some time but the actual move type and observables have now been tested.
There is a "spin flip" move type which alters an integer state on a randomly chosen molecule.  This flag flips between either "para" or "ortho".  After calculation of the rotational eigenstates, a molecular partition function is calculated by direct summation over states BUT only includes the states of the same symmetry (g for symmetric states for para, or u for antisymmetric states for ortho).  The spin flip move is accepted based upon Metropolis criteria that weights these molecular partition functions (in addition to the usual energy change).

Convergence to the classical limit of 3:1 ortho:para is a bit problematic since the number of eigenstates is quite small (for numerical efficiency, typically not summing over more than 10 energy levels) but for temperatures less than the characteristic rotational temperature it seems to work.

Also added a few generic things needed for compilation on MacOS